+++ R/output.R
@@ -387,7 +387,7 @@ wrap.recordedplot = function(x, options) {
   if (!file.exists(dirname(name)))
     dir.create(dirname(name), recursive = TRUE) # automatically creates dir for plots
   ## vectorize over dev, ext and dpi: save multiple versions of the plot
-  name.ext = mapply(save_plot,
+  name.ext = mapply(save_plot, width = options$fig.width, height = options$fig.height,
                     dev = options$dev, ext = options$fig.ext, dpi = options$dpi,
                     MoreArgs = list(plot = x, name = name, options = options),
                     SIMPLIFY = FALSE)[[1]]